Adding a double-escape mechanism to the source path to fix an issue in Windows.
Windows OSes use the escape prefix (backslash) instead of the forward slash as the path separator character like most sensible operating systems. This results in the plugin (correctly) interpreting the separator character as an escape sequence, and when we attempt to convert the path into a regexp, Clojure complains that the escape sequences are invalid. By double-escaping them, we prevent this from occuring.

This might end up introducing a breaking change if for some reason any users are inputting a regexp string as the 'source-dir' configuration parameter.